requirements.txt
asgiref==3.3.1
Django==3.0.8
djangorestframework==3.11.0
pytz==2021.1
sqlparse==0.4.1
